refactor: make numeric casts and sentinels explicit

+++ source/tools/finite/env/mixer.h
@@ -1,10 +1,11 @@
 #pragma once
+#include <cstddef>
 
 template<typename Scalar> class TensorsFinite;
 struct BondExpansionConfig;
 namespace tools::finite::env::internal {
 
     template<typename Scalar>
-    void run_expansion_term_mixer(TensorsFinite<Scalar> &tensors, long posP, long pos0, const BondExpansionConfig &bcfg);
+    void run_expansion_term_mixer(TensorsFinite<Scalar> &tensors, size_t posP, size_t pos0, const BondExpansionConfig &bcfg);
 
 }
